refactor(cli): split driver tslib helper checks
